use TAG in prod-65830 images; path other deployments to not fail on first deploy

+++ deploy/dev/dev-stop.sh
@@ -1,5 +1,9 @@
 #!/bin/bash
 
-echo "Stop running docker compose"
-cd A2rchi-dev/deploy/dev/
-docker compose -f dev-compose.yaml down
+if [ -z "$(ls -A A2rchi-dev/deploy/dev/)" ]; then
+    echo "Deployment directory is empty; skipping docker compose down"
+else
+    echo "Stop running docker compose"
+    cd A2rchi-dev/deploy/dev/
+    docker compose -f dev-compose.yaml down
+fi
